Dokel Edits Round 2
Few edits made that were suggested by Dokel. Changes include changing the order of imports for stylesheets and fonts, getting rid of an unnecessary href and id on line  20 and adding the id to the header section. Lastly removed a blank and incorrect size element in an image tag on line 109 which is now line 107 after some formatting.